refactor(#299): remove ResultUtil helper
