+++ lib/capybara-lib/src/main/java/dev/capylang/IOUtil.java
@@ -1,6 +1,8 @@
 package dev.capylang;
 
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.nio.file.AccessDeniedException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.CopyOption;
@@ -12,7 +14,10 @@ import java.nio.file.NoSuchFileException;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.NotDirectoryException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 public final class IOUtil {
     private IOUtil() {
@@ -22,7 +27,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(Files.readString(javaPath, StandardCharsets.UTF_8));
+            return success(Files.readString(javaPath, StandardCharsets.UTF_8));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("read_text", pathText(javaPath, path), e);
         }
@@ -32,7 +37,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(Files.readAllLines(javaPath, StandardCharsets.UTF_8));
+            return success(Files.readAllLines(javaPath, StandardCharsets.UTF_8));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("read_lines", pathText(javaPath, path), e);
         }
@@ -42,7 +47,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(toByteList(Files.readAllBytes(javaPath)));
+            return success(toByteList(Files.readAllBytes(javaPath)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("read_bytes", pathText(javaPath, path), e);
         }
@@ -60,7 +65,7 @@ public final class IOUtil {
                     StandardOpenOption.TRUNCATE_EXISTING,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(text);
+            return success(text);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("write_text", pathText(javaPath, path), e);
         }
@@ -78,7 +83,7 @@ public final class IOUtil {
                     StandardOpenOption.TRUNCATE_EXISTING,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(lines);
+            return success(lines);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("write_lines", pathText(javaPath, path), e);
         }
@@ -95,7 +100,7 @@ public final class IOUtil {
                     StandardOpenOption.TRUNCATE_EXISTING,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(bytes);
+            return success(bytes);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("write_bytes", pathText(javaPath, path), e);
         }
@@ -113,7 +118,7 @@ public final class IOUtil {
                     StandardOpenOption.APPEND,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(text);
+            return success(text);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("append_text", pathText(javaPath, path), e);
         }
@@ -131,7 +136,7 @@ public final class IOUtil {
                     StandardOpenOption.APPEND,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(lines);
+            return success(lines);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("append_lines", pathText(javaPath, path), e);
         }
@@ -148,7 +153,7 @@ public final class IOUtil {
                     StandardOpenOption.APPEND,
                     StandardOpenOption.WRITE
             );
-            return ResultUtil.success(bytes);
+            return success(bytes);
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("append_bytes", pathText(javaPath, path), e);
         }
@@ -170,7 +175,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(Files.size(javaPath));
+            return success(Files.size(javaPath));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("size", pathText(javaPath, path), e);
         }
@@ -180,7 +185,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(PathUtil.fromJavaPath(Files.createFile(javaPath)));
+            return success(PathUtil.fromJavaPath(Files.createFile(javaPath)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("create_file", pathText(javaPath, path), e);
         }
@@ -190,7 +195,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(PathUtil.fromJavaPath(Files.createDirectory(javaPath)));
+            return success(PathUtil.fromJavaPath(Files.createDirectory(javaPath)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("create_directory", pathText(javaPath, path), e);
         }
@@ -200,7 +205,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(PathUtil.fromJavaPath(Files.createDirectories(javaPath)));
+            return success(PathUtil.fromJavaPath(Files.createDirectories(javaPath)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("create_directories", pathText(javaPath, path), e);
         }
@@ -211,7 +216,7 @@ public final class IOUtil {
         try {
             javaPath = PathUtil.toJavaPath(path);
             try (var entries = Files.list(javaPath)) {
-                return ResultUtil.success(entries
+                return success(entries
                         .map(PathUtil::fromJavaPath)
                         .toList());
             }
@@ -224,7 +229,7 @@ public final class IOUtil {
         java.nio.file.Path javaPath = null;
         try {
             javaPath = PathUtil.toJavaPath(path);
-            return ResultUtil.success(Files.deleteIfExists(javaPath));
+            return success(Files.deleteIfExists(javaPath));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error("delete", pathText(javaPath, path), e);
         }
@@ -252,7 +257,7 @@ public final class IOUtil {
         try {
             javaSource = PathUtil.toJavaPath(source);
             javaTarget = PathUtil.toJavaPath(target);
-            return ResultUtil.success(PathUtil.fromJavaPath(Files.copy(javaSource, javaTarget, options)));
+            return success(PathUtil.fromJavaPath(Files.copy(javaSource, javaTarget, options)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error(operation, pathText(javaSource, source) + " -> " + pathText(javaTarget, target), e);
         }
@@ -264,7 +269,7 @@ public final class IOUtil {
         try {
             javaSource = PathUtil.toJavaPath(source);
             javaTarget = PathUtil.toJavaPath(target);
-            return ResultUtil.success(PathUtil.fromJavaPath(Files.move(javaSource, javaTarget, options)));
+            return success(PathUtil.fromJavaPath(Files.move(javaSource, javaTarget, options)));
         } catch (IOException | SecurityException | InvalidPathException e) {
             return error(operation, pathText(javaSource, source) + " -> " + pathText(javaTarget, target), e);
         }
@@ -276,7 +281,125 @@ public final class IOUtil {
 
     private static Object error(String operation, String path, Exception e) {
         var message = operation + " failed for `" + path + "`: " + e.getMessage();
-        return ResultUtil.error(kind(operation, e), message, e);
+        return resultError(kind(operation, e), message, e);
+    }
+
+    private static Object success(Object value) {
+        var result = new java.util.LinkedHashMap<String, Object>();
+        result.put("__type", "Success");
+        result.put("value", value);
+        return result;
+    }
+
+    private static Object resultError(String kind, String message, Throwable throwable) {
+        var stackTrace = stackTrace(throwable);
+        return errorFull(
+                kind,
+                message,
+                Optional.empty(),
+                sourceLocation(stackTrace),
+                stackTrace,
+                Optional.of(rawStack(throwable)),
+                cause(throwable),
+                suppressed(throwable)
+        );
+    }
+
+    private static Object errorFull(
+            String kind,
+            String message,
+            Optional<Object> details,
+            Optional<Object> location,
+            List<Object> stackTrace,
+            Optional<String> rawStack,
+            Optional<Object> cause,
+            List<Object> suppressed
+    ) {
+        return Map.ofEntries(
+                Map.entry("__type", "Error"),
+                Map.entry("kind", kind),
+                Map.entry("message", message),
+                Map.entry("details", details),
+                Map.entry("location", location),
+                Map.entry("stack_trace", stackTrace),
+                Map.entry("raw_stack", rawStack),
+                Map.entry("cause", cause),
+                Map.entry("suppressed", suppressed)
+        );
+    }
+
+    private static List<Object> stackTrace(Throwable throwable) {
+        var frames = new ArrayList<Object>();
+        for (var element : throwable.getStackTrace()) {
+            frames.add(stackFrame(element));
+        }
+        return List.copyOf(frames);
+    }
+
+    private static Object stackFrame(StackTraceElement element) {
+        return Map.ofEntries(
+                Map.entry("__type", "StackFrame"),
+                Map.entry("backend", "java"),
+                Map.entry("module", Optional.ofNullable(element.getModuleName())),
+                Map.entry("type_name", Optional.ofNullable(element.getClassName())),
+                Map.entry("function", Optional.ofNullable(element.getMethodName())),
+                Map.entry("file", Optional.ofNullable(element.getFileName())),
+                Map.entry("line", element.getLineNumber() >= 0 ? Optional.of(element.getLineNumber()) : Optional.empty()),
+                Map.entry("column", Optional.empty()),
+                Map.entry("end_line", Optional.empty()),
+                Map.entry("end_column", Optional.empty()),
+                Map.entry("source_line", Optional.empty()),
+                Map.entry("native", element.isNativeMethod()),
+                Map.entry("raw", element.toString())
+        );
+    }
+
+    private static Optional<Object> sourceLocation(List<Object> stackTrace) {
+        for (var frame : stackTrace) {
+            if (frame instanceof Map<?, ?> map && map.get("file") instanceof Optional<?> file && file.isPresent()) {
+                var line = map.get("line") instanceof Optional<?> lineOption && lineOption.isPresent()
+                        ? (Integer) lineOption.get()
+                        : 0;
+                return Optional.of(Map.ofEntries(
+                        Map.entry("__type", "SourceLocation"),
+                        Map.entry("file", String.valueOf(file.get())),
+                        Map.entry("line", line),
+                        Map.entry("column", Optional.empty()),
+                        Map.entry("end_line", Optional.empty()),
+                        Map.entry("end_column", Optional.empty())
+                ));
+            }
+        }
+        return Optional.empty();
+    }
+
+    private static Optional<Object> cause(Throwable throwable) {
+        var cause = throwable.getCause();
+        if (cause == null || cause == throwable) {
+            return Optional.empty();
+        }
+        return Optional.of(resultError(cause.getClass().getName(), safeMessage(cause), cause));
+    }
+
+    private static List<Object> suppressed(Throwable throwable) {
+        var suppressed = new ArrayList<Object>();
+        for (var item : throwable.getSuppressed()) {
+            suppressed.add(resultError(item.getClass().getName(), safeMessage(item), item));
+        }
+        return List.copyOf(suppressed);
+    }
+
+    private static String rawStack(Throwable throwable) {
+        var writer = new StringWriter();
+        throwable.printStackTrace(new PrintWriter(writer));
+        return writer.toString();
+    }
+
+    private static String safeMessage(Throwable throwable) {
+        var message = throwable.getMessage();
+        return message == null || message.isBlank()
+                ? throwable.getClass().getSimpleName()
+                : message;
     }
 
     private static String kind(String operation, Exception e) {
